pyproject.toml
[project]
authors = [{name = "Jonathan Moore", email = "jonathanmoorephd@gmail.com"}]
dependencies = []
description = "Advent of Code solutions"
name = "aoc"
requires-python = ">= 3.12"
version = "0.1.0"

[project.optional-dependencies]
test = ["pytest-line-profiler>=0.2.1"]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
aoc = { path = ".", editable = true }

[tool.pixi.tasks]

[tool.pixi.feature.test.dependencies]
ruff = "*"
mypy = "*"
pytest = "*"

[tool.pixi.feature.test.tasks]
test = "pytest"
fmt = "ruff format"
fix = "ruff format && ruff check --fix"

[tool.pixi.dependencies]
ipython = ">=8.31.0"
networkx = ">=3.4.2"
line_profiler = ">=4.1.3"

[tool.pixi.environments]
default = { solve-group = "default" }
test = { features = ["test"], solve-group = "default" }

[tool.pytest.ini_options]
markers = [
    "heavy: mark a test as heavy (slow or resource-intensive)"
]

